pyproject.toml
[project]
name = "drission-rod"
version = "0.1.0"
description = "drission-rod"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ddddocr>=1.5.6",
    "drissionpage>=4.1.1.2",
    "fastapi>=0.116.1",
    "grpcio>=1.74.0",
    "loguru>=0.7.3",
    "pydantic-settings>=2.10.1",
    "uvicorn>=0.35.0",
]

[tool.uv]
index-url = "https://mirrors.aliyun.com/pypi/simple/"

[dependency-groups]
dev = [
    "grpcio-tools>=1.74.0",
    "ruff>=0.13.0",
    "setuptools>=80.9.0",
]
